:zap: Minor refactoring and modernization of monaco-related code
